+++ lib/logging/brief.rb
@@ -88,8 +88,7 @@ class LoggingBrief < Logging
     else
       brief_results_final = "#{target} [#{status} #{status_code}] #{brief_results.join(', ')}"
     end
-    $semaphore.synchronize do
-      @f.puts brief_results_final
-    end
+    # Use buffered write for better performance
+    buffered_write(brief_results_final)
   end
 end